feat: moved POST to api/events
