+++ src/app/api/events/[eventID]/route.ts
@@ -1,6 +1,6 @@
 import { NextRequest, NextResponse } from "next/server";
 import connectDB from "@database/db";
-import EventSchema, { IEvent } from "@database/eventSchema";
+import EventSchema from "@database/eventSchema";
 
 type IParams = {
   params: {
@@ -22,38 +22,13 @@ export async function GET(req: NextRequest, { params }: IParams) {
   }
 }
 
-export async function POST(req: NextRequest) {
-  await connectDB();
-  try {
-    const { picture, alt, description, date, name, eventID }: IEvent =
-      await req.json();
-    if (!(picture && alt && description && date && name && eventID)) {
-      return NextResponse.json("Failed: Invalid Event", { status: 400 });
-    }
-
-    const newEvent = new EventSchema({
-      picture,
-      alt,
-      description,
-      date,
-      name,
-      eventID,
-    });
-    console.log(newEvent);
-    await newEvent.save();
-    return NextResponse.json("Success: Event uploaded", { status: 200 });
-  } catch (err) {
-    return NextResponse.json(`${err}`, { status: 400 });
-  }
-}
-
 export async function DELETE(req: NextRequest, { params }: IParams) {
   await connectDB();
   const { eventID } = params;
 
   try {
-    const deleted = await EventSchema.deleteOne({ eventID }).orFail();
-    return NextResponse.json(`Event deleted: ${deleted}`, { status: 200 });
+    await EventSchema.deleteOne({ eventID }).orFail();
+    return NextResponse.json(`Event deleted.`, { status: 200 });
   } catch (err) {
     return NextResponse.json(`Event not found. Error: ${err}`, { status: 400 });
   }
