+++ src/app/api/events/route.ts
@@ -1,6 +1,6 @@
 import { NextRequest, NextResponse } from "next/server";
 import connectDB from "@database/db";
-import EventSchema from "@database/eventSchema";
+import EventSchema, { IEvent } from "@database/eventSchema";
 
 export async function GET(req: NextRequest) {
   await connectDB();
@@ -12,5 +12,31 @@ export async function GET(req: NextRequest) {
     return NextResponse.json(`Events could not be found. Error: ${err}`, {
       status: 400,
     });
+  }
+}
+
+export async function POST(req: NextRequest) {
+  await connectDB();
+
+  try {
+    const { picture, alt, description, date, name, eventID }: IEvent =
+      await req.json();
+    if (!(picture && alt && description && date && name && eventID)) {
+      return NextResponse.json("Failed: Invalid Event", { status: 400 });
+    }
+
+    const newEvent = new EventSchema({
+      picture,
+      alt,
+      description,
+      date,
+      name,
+      eventID,
+    });
+    console.log(newEvent);
+    await newEvent.save();
+    return NextResponse.json("Success: Event uploaded", { status: 200 });
+  } catch (err) {
+    return NextResponse.json(`${err}`, { status: 400 });
   }
 }
